[REFACTOR] #26 - 패키징 정리
