+++ app/src/main/java/com/example/sobok_android/data/repository/notice/NoticeRepositoryImpl.kt
@@ -1,9 +1,9 @@
-package com.example.sobok_android.data.repository
+package com.example.sobok_android.data.repository.notice
 
 import com.example.sobok_android.data.datasource.notice.NoticeDataSource
 import com.example.sobok_android.data.mapper.NoticeMapper
 import com.example.sobok_android.domain.model.notice.NoticeListData
-import com.example.sobok_android.domain.repository.NoticeRepository
+import com.example.sobok_android.domain.repository.notice.NoticeRepository
 
 class NoticeRepositoryImpl(private val noticeDataSource : NoticeDataSource) : NoticeRepository {
     override suspend fun getNoticeList(): NoticeListData {
