+++ app/src/main/java/com/example/sobok_android/domain/repository/notice/NoticeRepository.kt
@@ -1,4 +1,4 @@
-package com.example.sobok_android.domain.repository
+package com.example.sobok_android.domain.repository.notice
 
 import com.example.sobok_android.data.model.response.notice.ResNoticeListData
 import com.example.sobok_android.domain.model.notice.NoticeListData
