+++ app/src/main/java/com/example/sobok_android/presentation/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.sobok_android.presentation.di
 
-import com.example.sobok_android.data.repository.NoticeRepositoryImpl
-import com.example.sobok_android.domain.repository.NoticeRepository
+import com.example.sobok_android.data.repository.notice.NoticeRepositoryImpl
+import com.example.sobok_android.domain.repository.notice.NoticeRepository
 import com.example.sobok_android.data.repository.pill.pilladd.PillAddRepositoryImpl
 import com.example.sobok_android.data.repository.share.request.ShareRequestRepositoryImpl
 import com.example.sobok_android.domain.repository.pill.pilladd.PillAddRepository
